FIX: status column type is short, do not use int

+++ scp-write-service-app/src/main/java/com/lastrix/scp/writeservice/dao/PostgreEnrolleeDao.java
@@ -73,7 +73,7 @@ public class PostgreEnrolleeDao implements EnrolleeDao {
                 (rs, rowNum) -> {
                     var r = new EnrolleSelectInfo();
                     r.setSpecId(UUID.fromString(rs.getString(1)));
-                    r.setStatus(rs.getInt(2));
+                    r.setStatus(rs.getShort(2));
                     r.setScore(rs.getInt(3));
                     r.setCreatedStamp(toInstantOrNull(rs, 4));
                     r.setConfirmedStamp(toInstantOrNull(rs, 5));
